BigDecimal + refactoring

+++ src/main/java/uk/org/carbonintensity/rest/contract/GenerationMix.java
@@ -2,13 +2,15 @@ package uk.org.carbonintensity.rest.contract;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 public class GenerationMix {
 
     @JsonProperty
     private String fuel;
 
     @JsonProperty
-    private double perc;
+    private BigDecimal perc;
 
     public String getFuel() {
         return fuel;
@@ -18,11 +20,11 @@ public class GenerationMix {
         this.fuel = fuel;
     }
 
-    public double getPerc() {
+    public BigDecimal getPerc() {
         return perc;
     }
 
-    public void setPerc(double perc) {
+    public void setPerc(BigDecimal perc) {
         this.perc = perc;
     }
 }
